Removed the DSL macros, More fixes related to memory.
I commented out stuff related to type caching, I may not be using it. Depends on whether I want to keep the parent member in the ast.

DSL macros were removed, I want to keep the macros minimum for the api.
The varadic macros for the body may be problomatic if there is some bs vendor's are doing with memory layouts of structs that are not necessarily POD but only have the assignment operator overloaded.
Worst case I'll have to remove them.

Memory should be fine now to start doing iterations on the array test.

Parser constructors have been on hold for a while. They'll problably not be done until sometime in June.